# docs/swagger.yaml
definitions:
  errormodel.Error:
    properties:
      error:
        $ref: '#/definitions/errormodel.ErrorDescriptor'
    type: object
  errormodel.ErrorDescriptor:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  gamemodel.Arg:
    properties:
      argument:
        type: string
      argumentseperator:
        type: string
      createdat:
        type: string
      defaultbool:
        type: boolean
      defaultfloat:
        type: number
      defaultint:
        type: integer
      defaultstring:
        type: string
      description:
        type: string
      enabled:
        type: boolean
      enumvalues:
        items:
          type: string
        type: array
      floatprecision:
        type: integer
      format:
        type: string
      maxfloat:
        type: number
      maxint:
        type: integer
      minfloat:
        type: number
      minint:
        type: integer
      orderindex:
        type: integer
      required:
        type: boolean
      role:
        type: string
    type: object
  gamemodel.Executable:
    properties:
      args:
        items:
          $ref: '#/definitions/gamemodel.Arg'
        type: array
      argumentseperator:
        type: string
      createdat:
        type: string
      format:
        type: string
      path:
        type: string
      requiresadmin:
        type: boolean
      role:
        type: string
    type: object
  gamemodel.Game:
    properties:
      createdat:
        type: string
      executables:
        items:
          $ref: '#/definitions/gamemodel.Executable'
        type: array
      name:
        type: string
      slug:
        type: string
    type: object
  gamemodel.UpsertGameExecutableArgRequest:
    properties:
      argument:
        type: string
      argumentseperator:
        type: string
      defaultbool:
        type: boolean
      defaultfloat:
        type: number
      defaultint:
        type: integer
      defaultstring:
        type: string
      description:
        type: string
      enabled:
        type: boolean
      enumvalues:
        items:
          type: string
        type: array
      floatprecision:
        type: integer
      format:
        type: string
      maxfloat:
        type: number
      maxint:
        type: integer
      minfloat:
        type: number
      minint:
        type: integer
      name:
        type: string
      orderindex:
        type: integer
      required:
        type: boolean
      role:
        type: string
    type: object
  gamemodel.UpsertGameExecutableRequest:
    properties:
      args:
        items:
          $ref: '#/definitions/gamemodel.UpsertGameExecutableArgRequest'
        type: array
      argumentseperator:
        type: string
      format:
        type: string
      path:
        type: string
      requiresadmin:
        type: boolean
      role:
        type: string
    type: object
  gamemodel.UpsertGameRequest:
    properties:
      executables:
        items:
          $ref: '#/definitions/gamemodel.UpsertGameExecutableRequest'
        type: array
      name:
        type: string
    type: object
  usermodel.UpsertUserRequest:
    properties:
      username:
        type: string
    type: object
  usermodel.User:
    properties:
      ipv4Address:
        type: string
      username:
        type: string
    type: object
externalDocs:
  description: Documentation
  url: https://github.com/seternate/go-lanty/blob/main/README.md
info:
  contact:
    email: seternate@gmail.com
    name: Levin Jeck
    url: https://github.com/seternate/go-lanty
  description: Lanty is a platform for managing and serving games for LAN parties
  license:
    name: License - MIT
    url: https://github.com/seternate/go-lanty/blob/main/LICENSE.md
  title: Lanty
  version: dev
paths:
  /games:
    get:
      description: Get all available Games
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/gamemodel.Game'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errormodel.Error'
      summary: Get all Games
      tags:
      - games
  /games/{slug}:
    delete:
      description: Delete a Game
      parameters:
      - description: Slug
        in: path
        name: slug
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errormodel.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errormodel.Error'
      summary: Delete a Game
      tags:
      - games
    get:
      description: Get a Game by its slug
      parameters:
      - description: Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gamemodel.Game'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errormodel.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errormodel.Error'
      summary: Get a Game
      tags:
      - games
    put:
      consumes:
      - application/json
      description: Create or update the given Game.
      parameters:
      - description: Slug
        in: path
        name: slug
        required: true
        type: string
      - description: Game to create or update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/gamemodel.UpsertGameRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Game created
          schema:
            $ref: '#/definitions/gamemodel.Game'
        "202":
          description: Game updated
          schema:
            $ref: '#/definitions/gamemodel.Game'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errormodel.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errormodel.Error'
      summary: Create or update a Game
      tags:
      - games
  /games/{slug}/blob:
    get:
      description: Get the blob for a Game
      parameters:
      - description: Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/octet-stream
      - ' application/json'
      responses:
        "200":
          description: Blob binary data
          headers:
            Content-Digest:
              description: 'Checksum (RFC 9530: algorithm=base64_checksum)'
              type: string
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errormodel.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errormodel.Error'
      summary: Get a Games blob
      tags:
      - games
    put:
      consumes:
      - application/octet-stream
      description: Updates a Games blob.
      parameters:
      - description: Slug
        in: path
        name: slug
        required: true
        type: string
      - description: 'Checksum (RFC 9530: algorithm=base64_checksum)'
        in: header
        name: Content-Digest
        required: true
        type: string
      - description: Blob binary data
        in: body
        name: request
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Blob created
        "202":
          description: Blob updated
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errormodel.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errormodel.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errormodel.Error'
      summary: Update a Games blob
      tags:
      - games
  /games/{slug}/icon:
    get:
      description: Get the icon of a Game
      parameters:
      - description: Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - image/*
      - ' application/json'
      responses:
        "200":
          description: Icon binary data
          headers:
            Content-Digest:
              description: 'Checksum (RFC 9530: algorithm=base64_checksum)'
              type: string
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errormodel.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errormodel.Error'
      summary: Get the icon of a Game
      tags:
      - games
    put:
      consumes:
      - image/*
      - ' multipart/form-data'
      description: |-
        Updates a Games icon.
        Accepts icon data in two formats:
        1) Raw binary data
        2) multipart/form-data with exactly one file. The Content-Digest header must match the checksum of the uploaded file data.
      parameters:
      - description: Slug
        in: path
        name: slug
        required: true
        type: string
      - description: 'Checksum (RFC 9530: algorithm=base64_checksum)'
        in: header
        name: Content-Digest
        required: true
        type: string
      - description: Icon binary data
        in: body
        name: request
        schema:
          type: string
      - description: Icon file (for multipart/form-data uploads). Exactly one file
          must be provided when using multipart format.
        in: formData
        name: file
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Icon created
        "202":
          description: Icon updated
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errormodel.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errormodel.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errormodel.Error'
      summary: Update a Games icon
      tags:
      - games
  /users:
    get:
      description: Get all available Users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/usermodel.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errormodel.Error'
      summary: Get all Users
      tags:
      - users
  /users/{ipv4Address}:
    delete:
      description: Delete a User
      parameters:
      - description: IPv4 Address
        in: path
        name: ipv4Address
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errormodel.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errormodel.Error'
      summary: Delete a User
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Create or update the given User.
      parameters:
      - description: IPv4 Address
        in: path
        name: ipv4Address
        required: true
        type: string
      - description: User to create or update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/usermodel.UpsertUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created
          schema:
            $ref: '#/definitions/usermodel.User'
        "202":
          description: User updated
          schema:
            $ref: '#/definitions/usermodel.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errormodel.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errormodel.Error'
      summary: Create or update a User
      tags:
      - users
swagger: "2.0"